+++ packages/grid/x-data-grid/src/components/base/GridBody.tsx
@@ -2,29 +2,8 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import { unstable_useEnhancedEffect as useEnhancedEffect } from '@mui/utils';
 import { useGridPrivateApiContext } from '../../hooks/utils/useGridPrivateApiContext';
-import { useGridSelector } from '../../hooks/utils/useGridSelector';
 import { GridMainContainer } from '../containers/GridMainContainer';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
-import {
-  gridColumnPositionsSelector,
-  gridColumnVisibilityModelSelector,
-  gridVisibleColumnDefinitionsSelector,
-} from '../../hooks/features/columns/gridColumnsSelector';
-import { gridFilterActiveItemsLookupSelector } from '../../hooks/features/filter/gridFilterSelector';
-import { gridSortColumnLookupSelector } from '../../hooks/features/sorting/gridSortingSelector';
-import {
-  gridTabIndexColumnHeaderSelector,
-  gridTabIndexCellSelector,
-  gridFocusColumnHeaderSelector,
-  unstable_gridTabIndexColumnGroupHeaderSelector,
-  unstable_gridFocusColumnGroupHeaderSelector,
-} from '../../hooks/features/focus/gridFocusStateSelector';
-import { gridDensityFactorSelector } from '../../hooks/features/density/densitySelector';
-import {
-  gridColumnGroupsHeaderMaxDepthSelector,
-  gridColumnGroupsHeaderStructureSelector,
-} from '../../hooks/features/columnGrouping/gridColumnGroupsSelector';
-import { gridColumnMenuSelector } from '../../hooks/features/columnMenu/columnMenuSelector';
 
 interface GridBodyProps {
   children?: React.ReactNode;
@@ -42,39 +21,6 @@ function GridBody(props: GridBodyProps) {
   const apiRef = useGridPrivateApiContext();
   const rootProps = useGridRootProps();
   const rootRef = React.useRef<HTMLDivElement>(null);
-
-  const visibleColumns = useGridSelector(apiRef, gridVisibleColumnDefinitionsSelector);
-  const filterColumnLookup = useGridSelector(apiRef, gridFilterActiveItemsLookupSelector);
-  const sortColumnLookup = useGridSelector(apiRef, gridSortColumnLookupSelector);
-  const columnPositions = useGridSelector(apiRef, gridColumnPositionsSelector);
-  const columnHeaderTabIndexState = useGridSelector(apiRef, gridTabIndexColumnHeaderSelector);
-  const cellTabIndexState = useGridSelector(apiRef, gridTabIndexCellSelector);
-  const columnGroupHeaderTabIndexState = useGridSelector(
-    apiRef,
-    unstable_gridTabIndexColumnGroupHeaderSelector,
-  );
-
-  const columnHeaderFocus = useGridSelector(apiRef, gridFocusColumnHeaderSelector);
-  const columnGroupHeaderFocus = useGridSelector(
-    apiRef,
-    unstable_gridFocusColumnGroupHeaderSelector,
-  );
-
-  const densityFactor = useGridSelector(apiRef, gridDensityFactorSelector);
-  const headerGroupingMaxDepth = useGridSelector(apiRef, gridColumnGroupsHeaderMaxDepthSelector);
-
-  const columnMenuState = useGridSelector(apiRef, gridColumnMenuSelector);
-  const columnVisibility = useGridSelector(apiRef, gridColumnVisibilityModelSelector);
-  const columnGroupsHeaderStructure = useGridSelector(
-    apiRef,
-    gridColumnGroupsHeaderStructureSelector,
-  );
-
-  const hasOtherElementInTabSequence = !(
-    columnGroupHeaderTabIndexState === null &&
-    columnHeaderTabIndexState === null &&
-    cellTabIndexState === null
-  );
 
   const [isVirtualizationDisabled, setIsVirtualizationDisabled] = React.useState(
     rootProps.disableVirtualization,
@@ -132,40 +78,17 @@ function GridBody(props: GridBodyProps) {
   apiRef.current.unstable_disableVirtualization = disableVirtualization;
   apiRef.current.unstable_enableVirtualization = enableVirtualization;
 
-  const columnHeadersRef = React.useRef<HTMLDivElement>(null);
-  const columnsContainerRef = React.useRef<HTMLDivElement>(null);
+  const hasDimensions = !!apiRef.current.getRootDimensions();
+
   const virtualScrollerRef = React.useRef<HTMLDivElement>(null);
 
   apiRef.current.register('private', {
-    columnHeadersContainerElementRef: columnsContainerRef,
-    columnHeadersElementRef: columnHeadersRef,
     virtualScrollerRef,
     mainElementRef: rootRef,
   });
 
-  const hasDimensions = !!apiRef.current.getRootDimensions();
-
   return (
     <GridMainContainer ref={rootRef}>
-      <rootProps.slots.columnHeaders
-        ref={columnsContainerRef}
-        innerRef={columnHeadersRef}
-        visibleColumns={visibleColumns}
-        filterColumnLookup={filterColumnLookup}
-        sortColumnLookup={sortColumnLookup}
-        columnPositions={columnPositions}
-        columnHeaderTabIndexState={columnHeaderTabIndexState}
-        columnGroupHeaderTabIndexState={columnGroupHeaderTabIndexState}
-        columnHeaderFocus={columnHeaderFocus}
-        columnGroupHeaderFocus={columnGroupHeaderFocus}
-        densityFactor={densityFactor}
-        headerGroupingMaxDepth={headerGroupingMaxDepth}
-        columnMenuState={columnMenuState}
-        columnVisibility={columnVisibility}
-        columnGroupsHeaderStructure={columnGroupsHeaderStructure}
-        hasOtherElementInTabSequence={hasOtherElementInTabSequence}
-        {...ColumnHeadersProps}
-      />
       {hasDimensions && (
         <VirtualScrollerComponent
           // The content is only rendered after dimensions are computed because
