+++ packages/grid/x-data-grid/src/components/columnHeaders/GridBaseColumnHeaders.tsx
@@ -23,8 +23,9 @@ const GridColumnHeadersRoot = styled('div', {
   slot: 'ColumnHeaders',
   overridesResolver: (props, styles) => styles.columnHeaders,
 })<{ ownerState: OwnerState }>({
-  position: 'relative',
-  overflow: 'hidden',
+  position: 'sticky',
+  top: 0,
+  zIndex: 2,
   display: 'flex',
   alignItems: 'center',
   boxSizing: 'border-box',
